Split planner into header and sourec
